update all entities, DTOs,Repos and services

+++ src/main/java/com/ammapickles/backend/dto/OrderDTO.java
@@ -3,6 +3,7 @@ package com.ammapickles.backend.dto;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -10,13 +11,16 @@ import java.util.List;
 @NoArgsConstructor
 @AllArgsConstructor
 public class OrderDTO {
-
-    private Long id;
-    private Long userId;
-    private List<OrderItemDTO> orderItems;
-    private double totalAmount;
-    private double deliveryCharge;
-    private String status;
-    private LocalDateTime orderDate;
-    private AddressDTO deliveryAddress;
+	
+	 private Long id;
+	    private Long userId; 
+	    private List<OrderItemDTO> orderItems;
+	    private BigDecimal totalAmount;
+	    
+	    
+	    private BigDecimal deliveryCharge;
+	    
+	    private String status;
+	    private LocalDateTime orderDate;
+	    private AddressDTO deliveryAddress;
 }
